tabla pivote

+++ database/migrations/2026_02_05_085808_create_telefonos_table.php
@@ -43,7 +43,7 @@ return new class extends Migration
                 ->cascadeOnUpdateon()
                 ->cascadeOnDelete();
 
-            $table->Integer('numero');
+            $table->bigInteger('numero');
         });
     }
